requirements.txt
python-fasthtml>=0.1.0
monsterui>=0.1.0
aiohttp>=3.8.0
requests>=2.28.0
python-dotenv>=1.0.0
uvicorn>=0.23.0
